Split the application code into different components
This commit applies the standard directory layout for a go application,
where the main.go is placed under the ./cmd directory and the rest of
the code, if not exposed publicly, is placed under ./internal.

It also breaks the application logic, placed all in the main.go before
this commit, into 4 different components.
- audit_writer: Writes in a file or to stderr any payload.
- metrics: Creates a metrics server.
- event_processor: It has the main Validator Webhook logic. It receives
  events, analyses them, writes them to a persistent medium using the
  audit_writer and creates a response.
- http_listener: Creates the server that receives POST calls from the
  K8S API. It forwards the requests to the event_processor.

One of the purposes of this refactor is to enable testing the
application using unittests. This can be done because the previous
components can only see their interfaces, so it's easy to use a mock
instead of the actual implementation when doing tests. The repo has
a "./mock" directory where all the mocks are placed.

There's a new CI job that executes all the unittests.

In ./docs there's a file called ForDevelopers.md that further explains
this new layout.

The Dockerfiles and the local-testing.sh script have been modified to
point to ./cmd/kube_audit as the new directory for the main.go file.

Regarding the end-to-end local test (the stuff in ./testing/locally),
the bash script has been modified to pick random ports when starting
the kube-audit-rest program. The go file has also been modified to
receive as a parameter the ports where it will send traffic.