Correct example redaction for secrets (#290)
Previously, this example configuration would only redact the data field in the configmap, but not redact the copy of the data field stored in the "kubectl.kubernetes.io/last-applied-configuration" annotation used by some Kubernetes API clients like kubectl.
This could lead to secrets being disclosed in the audit messages sent to the example elastic search.